# data/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Dietplanner API",
		"version": "1.0.0",
		"contact": {
			"name": "noTirT",
			"email": "tommanger55@gmail.com"
		}
	},
...
	"paths": {
...
		"/food": {
			"get": {
				"operationId": "getAllFood",
				"summary": "List of all Food items",
				"tags": ["Food"],
				"responses": {
					"200": {
						"description": "application/json",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FoodDBO"
								}
							}
						}
					}
				}
			},
			"post": {
				"operationId": "createNewFood",
				"summary": "Create a new Food item",
				"tags": ["Food"],
				"requestBody": {
					"description": "New Food item",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FoodDTO"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "The new Food item",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FoodDBO"
								}
							}
						}
					}
				}
			}
		},
		"/food/{foodId}": {
			"put": {
				"operationId": "updateFoodItem",
				"summary": "Update values of a Food item by its ID",
				"tags": ["Food"],
				"requestBody": {
					"description": "Updated item values",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FoodDTO"
							}
						}
					}
				},
				"parameters": [
					{
						"name": "foodId",
						"in": "path",
						"description": "ID of the Food item to be updated",
						"required": true,
						"schema": {
							"type": "string",
							"format": "ISO-8"
						}
					}
				],
				"responses": {
					"200": {
						"description": "The Food item was updated"
					}
				}
			},
			"delete": {
				"operationId": "deleteFoodById",
				"summary": "Delete a Food item by its ID",
				"tags": ["Food"],
				"parameters": [
					{
						"name": "foodId",
						"in": "path",
						"description": "ID of the Food item to be deleted",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "The Food item was deleted"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"FoodDBO": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"description": "Auto-generated id of the Food item",
						"example": "3"
					},
					"name": {
						"type": "string",
						"description": "Name of the Food item",
						"example": "Pizza"
					},
					"description": {
						"type": "string",
						"description": "Description of the food item",
						"required": false
					},
					"type": {
						"type": "string",
						"description": "Type of the Food item",
						"example": "laktosefrei"
					},
					"category": {
						"type": "string",
						"description": "Category of the Food item",
						"example": "Mittagessen"
					}
				}
			},
			"FoodDTO": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"description": "Name of the Food item",
						"example": "Pizza"
					},
					"description": {
						"type": "string",
						"description": "Description of the food item",
						"required": false
					},
					"type": {
						"type": "string",
						"description": "Type of the Food item",
						"example": "laktosefrei"
					},
					"category": {
						"type": "string",
						"description": "Category of the Food item",
						"example": "Mittagessen"
					}
				}
			},
...
		}
	}
}